requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
afinn=0.1=pypi_0
astor=0.7.1=py36_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.5.15=0
certifi=2019.6.16=py36_0
click=7.0=py36_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cupti=10.0.130=0
cython=0.29.11=pypi_0
darkflow=1.0.0=dev_0
dominate=2.3.5=pypi_0
flask=1.0.3=py36_0
flask-admin=1.5.3=pypi_0
flask-bootstrap=3.3.7.1=pypi_0
flask-login=0.4.1=py36_0
flask-sqlalchemy=2.4.0=py_0
flask-wtf=0.14.2=py36_1
freetype=2.9.1=h8a8886c_1
gast=0.2.2=py36_0
grpcio=1.16.1=py36hf8bcb03_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
intel-openmp=2019.4=243
itsdangerous=1.1.0=py36_0
jinja2=2.10.1=py36_0
joblib=0.13.2=py36_0
jpeg=9b=h024ee3a_2
kafka-python=1.4.6=py_0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.1.0=py_1
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36h7b6447c_0
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
mock=3.0.5=py36_0
ncurses=6.1=he6710b0_1
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
opencv3=3.1.0=py36_0
openssl=1.1.1c=h7b6447c_1
pandas=0.24.2=py36he6710b0_0
pillow=6.0.0=py36h34e0f95_0
pip=19.1.1=py36_0
protobuf=3.8.0=py36he6710b0_0
py4j=0.10.7=py36_0
pyspark=2.4.3=py_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
pyyaml=5.1=py36h7b6447c_0
readline=7.0=h7b6447c_5
scikit-learn=0.21.2=py36hd81dba3_0
scipy=1.2.1=py36h7c811a0_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
sqlalchemy=1.3.5=py36h7b6447c_0
sqlite=3.28.0=h7b6447c_0
tensorboard=1.13.1=py36hf484d3e_0
tensorflow-base=1.13.1=gpu_py36h8d69cac_0
tensorflow-estimator=1.13.0=py_0
tensorflow-gpu=1.13.1=pypi_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
visitor=0.1.3=pypi_0
werkzeug=0.15.4=py_0
wheel=0.33.4=py36_0
wtforms=2.2.1=pypi_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
